+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/apis/HubdleMinimalSourceSetConfigurableExtension.kt
@@ -5,6 +5,8 @@ import com.javiersc.kotlin.stdlib.capitalize
 import org.gradle.api.Action
 import org.gradle.api.NamedDomainObjectProvider
 import org.gradle.api.Project
+import org.jetbrains.kotlin.gradle.dsl.KotlinProjectExtension
+import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
 
 public abstract class HubdleMinimalSourceSetConfigurableExtension<T>(
     project: Project,
@@ -27,6 +29,13 @@ public abstract class HubdleMinimalSourceSetConfigurableExtension<T>(
     @HubdleDslMarker
     public fun test(action: Action<T>) {
         userConfigurable { test.configure(action) }
+    }
+
+    @HubdleDslMarker
+    public fun sourceSet(name: String, action: Action<KotlinSourceSet>) {
+        userConfigurable {
+            configure<KotlinProjectExtension> { sourceSets.named(name, action::execute) }
+        }
     }
 
     internal fun calculateName(name: String): String =