+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/android/application/HubdleKotlinAndroidApplicationExtension.kt
@@ -15,21 +15,19 @@ import com.javiersc.hubdle.extensions.kotlin.android._internal.configureAndroidA
 import com.javiersc.hubdle.extensions.kotlin.android.application.features.HubdleKotlinAndroidApplicationFeaturesExtension
 import com.javiersc.hubdle.extensions.kotlin.android.hubdleAndroid
 import com.javiersc.hubdle.extensions.kotlin.shared.HubdleKotlinMinimalSourceSetConfigurableExtension
+import com.javiersc.hubdle.extensions.shared.android.HubdleAndroidDelegateSharedApis
 import javax.inject.Inject
 import org.gradle.api.Action
 import org.gradle.api.Project
-import org.gradle.api.artifacts.Configuration
 import org.gradle.api.provider.Property
 import org.gradle.kotlin.dsl.configure
-import org.jetbrains.kotlin.gradle.dsl.KotlinProjectExtension
-import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
 
 @HubdleDslMarker
 public open class HubdleKotlinAndroidApplicationExtension
 @Inject
 constructor(
     project: Project,
-) : HubdleKotlinMinimalSourceSetConfigurableExtension(project) {
+) : HubdleKotlinMinimalSourceSetConfigurableExtension(project), HubdleAndroidDelegateSharedApis {
 
     override val project: Project
         get() = super.project
@@ -54,20 +52,6 @@ constructor(
     public val versionCode: Property<Int> = property { 1 }
 
     public val versionName: Property<String?> = property { "0.1.0" }
-
-    @HubdleDslMarker
-    public fun configuration(name: String, action: Action<Configuration>) {
-        userConfigurable {
-            configure<ApplicationExtension> { project.configurations.named(name, action::execute) }
-        }
-    }
-
-    @HubdleDslMarker
-    public fun sourceSet(name: String, action: Action<KotlinSourceSet>) {
-        userConfigurable {
-            configure<KotlinProjectExtension> { sourceSets.named(name, action::execute) }
-        }
-    }
 
     @HubdleDslMarker
     public fun android(action: Action<ApplicationExtension>) {