+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/multiplatform/targets/HubdleKotlinMultiplatformAndroidExtension.kt
@@ -6,13 +6,12 @@ import com.javiersc.hubdle.extensions._internal.ApplicablePlugin.Scope
 import com.javiersc.hubdle.extensions._internal.Configurable.Priority
 import com.javiersc.hubdle.extensions._internal.PluginId
 import com.javiersc.hubdle.extensions._internal.getHubdleExtension
-import com.javiersc.hubdle.extensions.apis.HubdleConfigurableExtension
 import com.javiersc.hubdle.extensions.apis.HubdleEnableableExtension
 import com.javiersc.hubdle.extensions.kotlin.android._internal.calculateAndroidNamespace
 import com.javiersc.hubdle.extensions.kotlin.android._internal.configureAndroidLibraryJavaVersion
-import com.javiersc.hubdle.extensions.kotlin.android.features.HubdleKotlinAndroidBuildFeaturesExtension
 import com.javiersc.hubdle.extensions.kotlin.multiplatform.hubdleKotlinMultiplatform
 import com.javiersc.hubdle.extensions.kotlin.shared.HubdleKotlinMinimalSourceSetConfigurableExtension
+import com.javiersc.hubdle.extensions.shared.android.HubdleAndroidDelegateSharedApis
 import javax.inject.Inject
 import org.gradle.api.Action
 import org.gradle.api.Project
@@ -26,7 +25,7 @@ public open class HubdleKotlinMultiplatformAndroidExtension
 @Inject
 constructor(
     project: Project,
-) : HubdleKotlinMinimalSourceSetConfigurableExtension(project) {
+) : HubdleKotlinMinimalSourceSetConfigurableExtension(project), HubdleAndroidDelegateSharedApis {
 
     override val project: Project
         get() = super.project
@@ -47,9 +46,6 @@ constructor(
     public val compileSdk: Property<Int> = property { 33 }
 
     public val targetSdk: Property<Int> = property { 33 }
-
-    public val buildFeatures: HubdleKotlinAndroidBuildFeaturesExtension
-        get() = getHubdleExtension()
 
     public val publishLibraryVariants: SetProperty<String> = setProperty { emptySet() }
 
@@ -111,6 +107,6 @@ constructor(
 internal val Project.hubdleKotlinMultiplatformAndroid: HubdleKotlinMultiplatformAndroidExtension
     get() = getHubdleExtension()
 
-internal val HubdleConfigurableExtension.hubdleKotlinMultiplatformAndroid:
+internal val HubdleEnableableExtension.hubdleKotlinMultiplatformAndroid:
     HubdleKotlinMultiplatformAndroidExtension
     get() = getHubdleExtension()